requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-generativeai>=0.8.3
pydantic>=2.0.0